Restricting users from deleting other user's rooms

+++ base/views.py
@@ -93,6 +93,10 @@ def updateRoom(request, pk):
 
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here !!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
